Move ApiVersions out of common to display
