+++ catalogue_api/api/src/main/scala/uk/ac/wellcome/platform/api/controllers/V1WorksController.scala
@@ -2,9 +2,9 @@ package uk.ac.wellcome.platform.api.controllers
 
 import com.twitter.inject.annotations.Flag
 import javax.inject.{Inject, Singleton}
+import uk.ac.wellcome.display.models.ApiVersions
 import uk.ac.wellcome.display.models.v1.DisplayWorkV1
 import uk.ac.wellcome.platform.api.services.WorksService
-import uk.ac.wellcome.versions.ApiVersions
 
 @Singleton
 class V1WorksController @Inject()(
