+++ catalogue_api/api/src/test/scala/uk/ac/wellcome/platform/api/works/v1/ApiV1WorksTest.scala
@@ -2,14 +2,9 @@ package uk.ac.wellcome.platform.api.works.v1
 
 import com.twitter.finagle.http.Status
 import com.twitter.finatra.http.EmbeddedHttpServer
-import uk.ac.wellcome.models.work.internal.{
-  DigitalLocation,
-  IdentifierSchemes,
-  License_CCBY,
-  SourceIdentifier
-}
+import uk.ac.wellcome.display.models.ApiVersions
+import uk.ac.wellcome.models.work.internal.{DigitalLocation, IdentifierSchemes, License_CCBY, SourceIdentifier}
 import uk.ac.wellcome.platform.api.works.ApiWorksTestBase
-import uk.ac.wellcome.versions.ApiVersions
 
 class ApiV1WorksTest extends ApiWorksTestBase {
