+++ data_api/snapshot_generator/src/main/scala/uk/ac/wellcome/platform/snapshot_generator/models/SnapshotJob.scala
@@ -1,6 +1,6 @@
 package uk.ac.wellcome.platform.snapshot_generator.models
 
-import uk.ac.wellcome.versions.ApiVersions
+import uk.ac.wellcome.display.models.ApiVersions
 
 case class SnapshotJob(publicBucketName: String,
                        publicObjectKey: String,
